Former-commit-id: 300b7032dec924f48795aeaef7563926a09097e7

+++ libraries/data/xmipp_image_base.h
@@ -436,7 +436,7 @@ public:
       *  If Image Object is read using readPreview method, movePointerTo only works when rescaling
       *  the image in X-Y plane only, but all slices must be read.
       */
-    virtual void movePointerTo(size_t select_slice = ALL_SLICES, size_t select_img = ALL_IMAGES) = 0;
+    virtual void movePointerTo(int select_slice = ALL_SLICES, size_t select_img = ALL_IMAGES) = 0;
 
     /* Return the datatype of the current image object */
     virtual DataType myT() const = 0;
